Add files via upload
Dijkstra algorithm to find the shortest path using adjacency list. 
Weight of the shortest path as well as the shortest path is printed. 
Min-Heap based priority queue is used to determine the shortest path. 
A vector array indexed by the vertex number is used to store the parent of any vertex which is traversed in the shortest path to reach the said vertex from a given source. 

example is assembled using code snippets from geeksforgeeks.org